Module colstodian::color

Contains color types and helper functions.

Structs

Color
A strongly typed color, parameterized by a color space and state.

ColorAlpha
A strongly typed color with an alpha channel, parameterized by a color space, state, and alpha state.

ColorU8
An encoded color, 8-bit per component, 24-bit total.

ColorU8Alpha
An encoded color with alpha, 8-bit per component, 32-bit total.

Functions

acescg
Creates a Color in the AcesCg color space in the St State with components r, g, and b.

linear_srgb
Creates a Color in the LinearSrgb color space in the St State with components r, g, and b.

linear_srgba
Creates a ColorAlpha in the LinearSrgb color space with components r, g, b, and a

srgb
Creates a Color in the EncodedSrgb color space with components r, g, and b.

srgb_u8
Creates a Color in the EncodedSrgb color space with components r, g, and b.

srgba
Creates a ColorAlpha in the EncodedSrgb color space with components r, g, b, and a.

srgba_u8
Creates a ColorU8Alpha in the EncodedSrgb color space with components r, g, b, and a.