Struct colstodian::color::Color

#[repr(transparent)]
pub struct Color<Spc, St> {
    pub raw: Vec3,
    // some fields omitted
}

A strongly typed color, parameterized by a color space and state.

See crate-level docs as well as ColorSpace and State for more.

Fields
raw: Vec3

The raw values of the color. Be careful when modifying this directly.

Implementations

impl<Spc, St> Color<Spc, St>

pub fn new(el1: f32, el2: f32, el3: f32) -> Self

Creates a Color with the internal color elements el1, el2, el3.

pub fn splat(el: f32) -> Self

Creates a Color with the internal color elements all set to el.

pub const fn from_raw(raw: Vec3) -> Self

Creates a Color with raw values contained in raw.

pub fn max_element(self) -> f32

Get the maximum element of self

pub fn min_element(self) -> f32

Get the minimum element of self

pub const ZERO: Self

pub const ONE: Self

impl<Spc: WorkingColorSpace, St> Color<Spc, St>

pub fn saturate(self) -> Self

Clamp the raw element values of self in the range [0..1]

impl<SrcSpace: ColorSpace, St: State> Color<SrcSpace, St>

pub fn convert<DstSpace: ConvertFromRaw<SrcSpace>>(self) -> Color<DstSpace, St>

Converts self from one color space to another while retaining the same State.

Be careful when converting between nonlinear color spaces while in Scene state as the nonlinear transform functions for some color spaces are only defined within a specific dynamic range.

pub fn convert_to<Query>(self) -> Color<Query::DstSpace, St> where
Query: ColorConversionQuery<SrcSpace, St>,

Converts self from one color space to another while retaining the same State.

This works the same as convert except there is only one type parameter, the “Query”.

The query is meant to be one of:

A ColorSpace

A Color (in which case it will be converted to that color’s space).

This query is slightly more generic than the ones on convert, which means that the Rust type system is usually not able to infer the query without you explicitly giving one.

This can be useful in conjunction with defined type aliases for predefined Color types.

pub fn cast_space<DstSpace: ColorSpace>(self) -> Color<DstSpace, St>

Interprets this color as DstSpace. This assumes you have done an external computation/conversion such that this cast is valid.

pub fn cast_state<DstSt>(self) -> Color<SrcSpace, DstSt>

Changes this color’s State. This assumes that you have done some kind of conversion externally, or that the proper conversion is simply a noop.

pub fn cast<DstSpace: ColorSpace, DstSt: State>(self) -> Color<DstSpace, DstSt>

Interprets this color as DstSpace and DstState. This assumes you have done an external computation/conversion such that this cast is valid.

impl<SrcSpace: EncodedColorSpace, St: State> Color<SrcSpace, St>

pub fn decode(self) -> Color<SrcSpace::DecodedSpace, St>

Decode self into its decoded (working) color space, which allows many more operations to be performed.

impl<SrcSpace: ColorSpace, St: State> Color<SrcSpace, St>

pub fn linearize(self) -> Color<SrcSpace::LinearSpace, St>

Convert self into the closest linear color space.

If self is already in a linear color space, this is a no-op.

impl<Spc: LinearColorSpace, SrcSt> Color<Spc, SrcSt>

pub fn convert_state<DstSt, F>(
self,
conversion_function: F
) -> Color<Spc, DstSt> where
F: FnOnce(Vec3) -> Vec3,

Converts this color from one state to another.

This conversion is usecase and even instance dependent. For example, converting a material’s emissive texture value, a Display-referred color, to a Scene-referred color might take the form of a multiplication which scales the power of said emission into Scene-referred irradiance. On the other hand, converting a final Scene-referred color to a Display-referred color should be done with some kind of tonemapping operator. For more, see the tonemap module.

Note that the conversion function gives a raw color value, as the state of the color during the intermediate steps of the conversion is not really well defined. Therefore it’s easier to just work on the raw values without type safety.

impl<Spc: WorkingColorSpace, St> Color<Spc, St>

pub fn blend(self, other: Color<Spc, St>, factor: f32) -> Color<Spc, St>

Blend self’s color values with the color values from other with linear interpolation. If factor is > 1.0, results may not be sensical.

impl<Spc: ColorSpace, St> Color<Spc, St>

pub fn with_alpha(self, alpha: f32) -> ColorAlpha<Spc, St, Separate>

Converts self to a ColorAlpha with Separate alpha state by adding a component. This is probably what you want.

pub fn to_alpha_opaque_premultiplied(self) -> ColorAlpha<Spc, St, Premultiplied>

Converts self to a ColorAlpha with fully opaque alpha (1.0) in premultiplied alpha state.

pub fn with_alpha_state<A: AlphaState>(
self,
alpha: f32
) -> ColorAlpha<Spc, St, A>

Converts self to a ColorAlpha with specified AlphaState by adding an alpha component. Make sure you choose the correct alpha state! If you’re not sure, you probably want Color::with_alpha.

impl<Spc: AsU8, St> Color<Spc, St>

pub fn to_u8(self) -> ColorU8<Spc, St>

Convert self to a [u8; 3]. All components of self will be clamped to range [0..1].

pub fn from_u8(col: ColorU8<Spc, St>) -> Color<Spc, St>

Decode a [u8; 4] into a Color with specified space and state.