Module colstodian::component_structs

Structs that act as bags of named components which Colors of different color spaces may be Dereferenced to in order to gain more appropriate dot syntax for that color space.

For example, a Color in the ICtCpPQ color space can be Deref’d to ICtCp, allowing you to do things like color.i or color.ct.

Structs

ColAlpha
A bag of components with an alpha channel. Some ColorAlphas will Deref to this so that you can use .alpha and .col.{component_name}.

ICtCp
A bag of components with names I (Intensity), Ct (Chroma-Tritan), Cp (Chroma-Protan). Some Colors will Deref/DerefMut to this struct so that you can access their components with dot-syntax.

LCh
A bag of components with names L (Luminance), C (chroma), h (hue). Some Colors with XYZ color spaces will Deref/DerefMut to this struct so that you can access their components with dot-syntax.

Lab
A bag of components with names L (Luminance), a (green-red chroma), b (blue-yellow chroma). Some Colors with XYZ color spaces will Deref/DerefMut to this struct so that you can access their components with dot-syntax.

Rgb
A bag of components with names R, G, B. Some Colors with RGB color spaces will Deref/DerefMut to this struct so that you can access their components with dot-syntax.

Xyz
A bag of components with names X, Y, Z. Some Colors with XYZ color spaces will Deref/DerefMut to this struct so that you can access their components with dot-syntax.