Module colstodian::details::traits

The traits which form the backbone of this crate.

Traits

AlphaOver

Implemented by color encodings that can do alpha compositing.

ColorEncoding

A type that implements ColorEncoding represents a collection of metadata that together defines how a color’s data is stored and what the values of that data actually mean.

ColorRepr

Implemented by the raw data representation of a color encoding.

ComponentStructFor

Implementing this trait for a struct marks that it is safe to pointer cast Repr as Self.

ConvertFrom

A trait that marks Self as being a color encoding which is able to be directly converted from SrcEnc, as well as allowing some hooks to perform extra mapping during the conversion if necessary. This is the trait that unlocks the .convert::<E> method on Color.

LinearColorSpace

A type that implements LinearColorSpace represents a color space which can be defined by a linear transformation only (i.e. a 3x3 matrix multiplication) from the CIE XYZ color space.

LinearConvertFromRaw

Performs the raw conversion from the LinearColorSpace represented by SrcSpc to the LinearColorSpace represented by Self.

LinearInterpolate

Implemented by color encodings which can perform linear interpolation between colors. The interpolation is not necessarily perceptually-linear, it is just linear within the given encoding.

PerceptualEncoding

Implemented by color encodings which are designed to be perceptually-uniform. In general these encodings will produce more visually pleasing results when blending between colors (for example, creating gradients) in many situations. However they are certainly not, silver bullets, and often don’t fully deliver on the promise of perceptual uniformity.

Saturate

Implemented by color encodings that can perform saturate-style clamping.

WorkingEncoding

Marks a type as representing a color encoding in which it makes sense to be able to perform mathematical operations on the contained color values directly.