Trait colstodian::details::traits::ColorEncoding

pub trait ColorEncoding: Sized + 'static {
    type Repr: ColorRepr;
    type ComponentStruct: ComponentStructFor<Self::Repr>;
    type LinearSpace: LinearColorSpace;

    const NAME: &'static str;

    // Required methods
    fn src_transform_raw(repr: Self::Repr) -> (Vec3, f32);
    fn dst_transform_raw(raw: Vec3, alpha: f32) -> Self::Repr;
}

A type that implements ColorEncoding represents a collection of metadata that together defines how a color’s data is stored and what the values of that data actually mean.

You can see a list of all the built-in color encodings in crate::details::encodings.

Required Associated Types

type Repr: ColorRepr

The raw data representation used by this encoding.

You can see a list of all the reprs used by the built-in encodings in crate::details::reprs.

type ComponentStruct: ComponentStructFor<Self::Repr>

The ‘bag of components’ this color encoding uses.

You can see a list of all the built-in component structs in crate::details::component_structs.

type LinearSpace: LinearColorSpace

The LinearColorSpace used by this encoding.

You can see all the built-in linear spaces in crate::details::linear_spaces

Required Associated Constants

const NAME: &'static str

Used in Debug and Display implementations.

Required Methods

fn src_transform_raw(repr: Self::Repr) -> (Vec3, f32)

Convert from Self::Repr to a glam::Vec3 in the Self::LinearSpace color space and a separate (not pre-multiplied) alpha component. If this encoding does not have alpha, return 1.0.

fn dst_transform_raw(raw: Vec3, alpha: f32) -> Self::Repr

Convert from a glam::Vec3 in Self::LinearSpace and separate alpha component to a Self::Repr fully encoded in Self’s color encoding. If this encoding does not have alpha, you can disregard it.