Trait colstodian::details::traits::ColorRepr

pub trait ColorRepr: Sized + Clone + Copy + 'static {
    type Element: Sized + Clone + Copy + 'static;
}

Implemented by the raw data representation of a color encoding.

You can see a list of all the reprs used by the built-in encodings in crate::details::reprs.

Required Associated Types

type Element: Sized + Clone + Copy + 'static

The type of a single element of this repr