Trait colstodian::details::traits::ComponentStructFor

pub unsafe trait ComponentStructFor<Repr: ColorRepr>: Sized + Clone + Copy + 'static {
    // Required methods
    fn cast(repr: &Repr) -> &Self;
    fn cast_mut(repr: &mut Repr) -> &mut Self;
}

Implementing this trait for a struct marks that it is safe to pointer cast Repr as Self.

Safety

In order to implement this trait, it must be safe to perform the casts implied by the cast and cast_mut functions.

Required Methods

fn cast(repr: &Repr) -> &Self

fn cast_mut(repr: &mut Repr) -> &mut Self