Trait colstodian::details::traits::ConvertFrom

pub trait ConvertFrom<SrcEnc>where
    SrcEnc: ColorEncoding,
    Self: ColorEncoding,
    Self::LinearSpace: LinearConvertFromRaw<SrcEnc::LinearSpace>,{
    // Provided method
    fn map_src(_src: &mut SrcEnc::Repr) { ... }
}

A trait that marks Self as being a color encoding which is able to be directly converted from SrcEnc, as well as allowing some hooks to perform extra mapping during the conversion if necessary. This is the trait that unlocks the .convert::<E> method on Color.

In order to be able to convert from Color<EncodingA> to Color<EncodingB>, EncodingB must implement ConvertFrom<EncodingA>.

If this trait is not implemented for a pair of encodings, then a direct conversion without input or choice from the user is not possible, and a conversion between the encodings will need to be performed manually or in more than one step.

Provided Methods

fn map_src(_src: &mut SrcEnc::Repr)

If required or desired, perform a mapping of some kind to the input before it undergoes its source transform. This may be desirable to perform some form of gamut mapping if the src encoding has a larger size of representable colors than the dst encoding.