Trait colstodian::details::traits::LinearColorSpace

pub trait LinearColorSpace {
    const PRIMARIES: RGBPrimaries;
    const WHITE_POINT: WhitePoint;
}

A type that implements LinearColorSpace represents a color space which can be defined by a linear transformation only (i.e. a 3x3 matrix multiplication) from the CIE XYZ color space.

A linear color space is defined by the combination of a set of Primaries and a White Point.

You can see all the built-in linear spaces in crate::details::linear_spaces

Required Associated Constants

const PRIMARIES: RGBPrimaries

const WHITE_POINT: WhitePoint