Trait colstodian::details::traits::LinearInterpolate

pub trait LinearInterpolatewhere
    Self: ColorEncoding + WorkingEncoding,{
    // Required method
    fn lerp(from: Color<Self>, to: Color<Self>, factor: f32) -> Color<Self>;
}

Implemented by color encodings which can perform linear interpolation between colors. The interpolation is not necessarily perceptually-linear, it is just linear within the given encoding.

Required Methods

fn lerp(from: Color<Self>, to: Color<Self>, factor: f32) -> Color<Self>