Module colstodian::spaces

Types representing different color spaces.

For more information, see the documentation for the corresponding dynamic color space (you can figure out the corresponding dynamic color space by looking at the implementation of the ColorSpace trait on a specific color space struct).

Modules

dynamic_spaces
Color spaces defined as data.

Structs

Aces2065
A type representing the ACES 2065-1 color space.

AcesCg
A type representing the ACEScg color space.

Bt2020
A type representing the BT.2020 color space (equivalent to the linear BT.2100 color space).

CieXYZ
A type representing the reference XYZ color space.

DisplayP3
A type representing the Apple Display P3 color space.

DynamicColorSpace
A color space defined in data by its Primaries, white point, and an optional invertible transform function.

EncodedAcesCgSrgb
A type representing the ACEScg color space encoded with the sRGB transfer functions.

EncodedBt2020
A type representing the encoded BT.2020 color space (with BT.2020 OETF applied).

EncodedBt2100PQ
A type representing the encoded BT.2100 color space (with inverse PQ EOTF applied).

EncodedDisplayP3
A type representing the encoded Display P3 color space (with sRGB OETF applied).

EncodedSrgb
A type representing the encoded sRGB colorspace.

ICtCpPQ
A type representing the ICtCp color space with PQ (Perceptual Quantizer) transfer functions.

LinearSrgb
A type representing the linear sRGB color space.

Oklab
A type representing the Oklab color space.

Oklch
A type representing the Oklch color space.

Type Definitions

Bt2100
A type representing the BT.2100 color space (equivalent to the unencoded BT.2020 color space).