Struct colstodian::Color

pub struct Color<E: ColorEncoding> {
    pub repr: E::Repr,
}

A strongly typed color, parameterized by a ColorEncoding.

Color is a unified type that defines a color in any ColorEncoding. The ColorEncoding defines a bunch of different properties about how the color values are stored and what those values actually mean. For example, Color<SrgbU8> is a color with red, green, and blue values that vary from 0-255 and the meaning of those values is defined by the full sRGB color encoding standard. The most common and standard color encodings are exposed in the basic_encodings module.

To create a new Color value, see the list of constructor helpers in the docs below.

There are several ways to work with an existing color. First, you can always access the raw data in the encoding’s Repr directly by accessing col.repr. You can also always access the individual named color components through dot-syntax because of the Deref and DerefMut impls to the encoding’s ComponentStruct. For example, in an RGB color space, you can access the components with .r, .g, and .b.

let col: Color<SrgbU8> = Color::srgb_u8(100u8, 105u8, 220u8);

assert_eq!(col.repr, [100u8, 105u8, 220u8]);

let mut col2: Color<LinearSrgb> = Color::linear_srgb(0.5, 1.0, 0.25);

assert_eq!(col2.repr, Vec3::new(0.5, 1.0, 0.25));

col2.b = 0.75;

assert_eq!(col2.r, 0.5);
assert_eq!(col2.g, 1.0);
assert_eq!(col2.b, 0.75);

In order to do math on color types without accessing the underlying repr directly, you’ll need to be in a WorkingEncoding, which is a trait implemented by encodings that support doing such math operations well.

You can convert between color encodings using the .convert::<E>() method.

Basic Conversion Example

Here we construct two colors in different ways, convert them both to LinearSrgb to work with them, and then convert the resul to SrgbU8 which can be passed on to be displayed in an image.

use colstodian::Color;
use colstodian::basic_encodings::{SrgbU8, LinearSrgb};

let color1 = Color::srgb_u8(102, 54, 220);
let color2 = Color::srgb_f32(0.5, 0.8, 0.1);

let color1_working = color1.convert::<LinearSrgb>();
let color2_working = color2.convert::<LinearSrgb>();

let result_working = color1_working * 0.5 + color2_working;

let output = result_working.convert::<SrgbU8>();

assert_eq!(output, Color::srgb_u8(144, 206, 163));

Fields
repr: E::Repr

The raw values of the color. Be careful when modifying this directly.

Implementations

impl Color<SrgbU8>

pub const fn srgb_u8(r: u8, g: u8, b: u8) -> Self

Create a Color in the SrgbU8 encoding.

If you don’t know what you’re doing and you have RGB values from a color picker that vary from 0-255, use this.

If you’re not sure, see the SrgbU8 encoding docs for more info.

impl Color<SrgbF32>

pub const fn srgb_f32(r: f32, g: f32, b: f32) -> Self

Create a Color in the SrgbF32 encoding.

If you don’t know what you’re doing and you have RGB values from a color picker that vary from 0.0..=1.0, use this.

If you’re not sure, see the SrgbF32 encoding docs for more info.

impl Color<SrgbAU8>

pub const fn srgba_u8(r: u8, g: u8, b: u8, a: u8) -> Self

Create a Color in the SrgbAU8 encoding.

If you don’t know what you’re doing and you have RGBA values from a color picker that vary from 0-255, use this.

If you’re not sure, see the SrgbAU8 encoding docs for more info.

impl Color<SrgbAF32>

pub const fn srgba_f32(r: f32, g: f32, b: f32, a: f32) -> Self

Create a Color in the SrgbAF32 encoding.

If you don’t know what you’re doing and you have RGB values from a color picker that vary from 0.0..=1.0, use this.

If you’re not sure, see the SrgbAF32 encoding docs for more info.

impl Color<LinearSrgb>

pub fn linear_srgb(r: f32, g: f32, b: f32) -> Self

Create a Color in the LinearSrgb encoding.

If you’re not sure, you should probably use Color::srgb_f32 instead. See the LinearSrgb encoding docs for more info.

impl Color<LinearSrgbA>

pub fn linear_srgba(r: f32, g: f32, b: f32, a: f32) -> Self

Create a Color in the LinearSrgbA encoding.

If you’re not sure, you should probably use Color::srgba_f32 instead. See the LinearSrgbA encoding docs for more info.

impl Color<LinearSrgbAPremultiplied>

pub fn linear_srgba_premultiplied(r: f32, g: f32, b: f32, a: f32) -> Self

Create a Color in the LinearSrgbAPremultiplied encoding.

“Premultiplied” alpha means that the value of the color components has been multiplied by the alpha component. This operation is unintuitive when specifying color values, but it is the “most correct” way to store color values with an alpha component when performing operations like blending and compositing on them.

If you’re not sure, see the LinearSrgbA encoding docs for more info.

impl Color<Oklab>

pub fn oklab(l: f32, a: f32, b: f32) -> Self

Create a Color in the Oklab color encoding.

This is fairly rare, it would be more common to specify colors in another color encoding like SrgbU8 and then convert them to Oklab to blend them together.

impl<E: ColorEncoding> Color<E>

pub const fn from_repr(repr: E::Repr) -> Self

Creates a Color from the raw data representation of the specified color encoding.

impl<SrcEnc: ColorEncoding> Color<SrcEnc>

pub fn convert<DstEnc>(self) -> Color<DstEnc>where DstEnc: ColorEncoding + ConvertFrom<SrcEnc>, DstEnc::LinearSpace: LinearConvertFromRaw<SrcEnc::LinearSpace>,

Converts self from one color encoding to another.

In order to be able to convert from EncodingA to EncodingB, EncodingB must implement ConvertFrom<EncodingA>.

If that trait is not implemented for a pair of encodings, then a direct conversion without input or choice from the user is not possible, and a conversion between the encodings will need to be performed manually or in more than one step.

If you are able to convert from EncodingA to EncodingB, then you can also use a Color<EncodingA> anywhere you need a type that implements ColorInto<Color<EncodingB>>!

Example

let grey_f32 = Color::srgb_f32(0.5, 0.5, 0.5);
let grey_u8 = Color::srgb_u8(127, 127, 127);

assert_eq_eps!(grey_f32.convert::<SrgbU8>(), grey_u8, 0);

let col = Color::srgb_u8(102, 51, 153);
let col_linear_srgb = col.convert::<LinearSrgb>();

assert_eq_eps!(col_linear_srgb, Color::linear_srgb(0.13287, 0.0331, 0.31855), 0.0001);

pub fn cast<DstEnc: ColorEncoding<Repr = SrcEnc::Repr>>(self) -> Color<DstEnc>

Interprets this color as DstEnc. Requires that DstEnc’s ColorEncoding::Repr is the same as self’s.

Using this method assumes you have done an external computation/conversion such that this cast is valid.

impl<E: ColorEncoding + Saturate> Color<E>

pub fn saturate(self) -> Self

Clamp the raw element values of self within the current color encoding’s valid range of values.

impl<E> Color<E>where E: ColorEncoding + AlphaOver,

pub fn alpha_over(self, under: Self) -> Color<E>

Alpha-composite self over under.

impl<E> Color<E>where E: ColorEncoding + PerceptualEncoding + LinearInterpolate, E::Repr: Add<Output = E::Repr> + Sub<Output = E::Repr> + Mul<f32, Output = E::Repr>,

pub fn perceptual_blend(self, other: Color<E>, factor: f32) -> Color<E>

Blend self’s color values with the color values from other with perceptually-linear interpolation.

factor ranges from [0..=1.0]. If factor is > 1.0, results may not be sensical.

impl<E> Color<E>where E: ColorEncoding + LinearInterpolate, E::Repr: Add<Output = E::Repr> + Sub<Output = E::Repr> + Mul<f32, Output = E::Repr>,

pub fn lerp(self, other: Self, factor: f32) -> Self

Linearly interpolate from self’s value to other’s value. Not guaranteed to be perceptually linear or pleasing!

If you want a better way to blend colors in a perceptually pleasing way, see Color::perceptual_blend, which requires that the color encoding is a PerceptualEncoding.

factor ranges from [0..=1.0]. If factor is > 1.0, results may not be sensical.