Struct colstodian::ColorU8Alpha

#[repr(C)]
pub struct ColorU8Alpha<Spc, St, A> {
    pub raw: [u8; 4],
    // some fields omitted
}

An encoded color with alpha, 8-bit per component, 32-bit total.

This should only be used when space is an issue, i.e. when compressing data. Otherwise prefer a ColorAlpha.

Fields
raw: [u8; 4]
Implementations

impl<Spc, St, A> ColorU8Alpha<Spc, St, A>

pub fn new(x: u8, y: u8, z: u8, w: u8) -> Self

pub fn from_raw(raw: [u8; 4]) -> Self

impl<Spc: AsU8, St, A> ColorU8Alpha<Spc, St, A>

pub fn from_f32(col: ColorAlpha<Spc, St, A>) -> ColorU8Alpha<Spc, St, A>

Convert self to a ColorU8Alpha of identical type. All components of self will be clamped to be in range [0..1].

pub fn to_f32(self) -> ColorAlpha<Spc, St, A>

Decode a ColorU8Alpha into a ColorAlpha of identical type.