Module colstodian::tonemap

Contains tonemappers, useful for mapping scene-referred HDR values into display-referred values within the concrete dynamic range of a specific display.

Structs

BakedLottesTonemapperParams

LottesTonemapper
See this talk by Timothy Lottes https://www.gdcvault.com/play/1023512/Advanced-Graphics-Techniques-Tutorial-Day and associated slides https://gpuopen.com/wp-content/uploads/2016/03/GdcVdrLottes.pdf

LottesTonemapperParams
Parameters for the LottesTonemapper

PerceptualTonemapper
A neutral, perceptual tonemapper based on tonemapping work in Frostbite as well as my own and Tomasz Stachowiak’s research on the subject. Should be able to be extended for HDR use in the future.

PerceptualTonemapperParams

Traits

Tonemapper
Performs tonemapping on a given input color.