Trait colstodian::AlphaState

pub trait AlphaState where
    Self: Default + ConvertToAlphaRaw<Separate> + ConvertToAlphaRaw<Premultiplied> + ConvertFromAlphaRaw<Separate> + ConvertFromAlphaRaw<Premultiplied>, {
    const STATE: DynamicAlphaState;
}

A type that implements this trait represents a color’s alpha state.

A color can either have a Separate alpha channel or have been pre-multiplied with its alpha channel and so have Premultiplied alpha.

Associated Constants

const STATE: DynamicAlphaState