Trait colstodian::traits::ColorConversionQuery

pub trait ColorConversionQuery<SrcSpace: ColorSpace, St: State> {
    type DstSpace: ConvertFromRaw<SrcSpace>;
}

A “conversion query” for a Color.

A type that implements this trait is able to be used as the type parameter for Color::convert.

The types that implement this trait are:

ColorSpace types

Color types (in which case it will be converted to that color’s space)

Associated Types

type DstSpace: ConvertFromRaw<SrcSpace>