Trait colstodian::traits::ConvertFromRaw

pub trait ConvertFromRaw<SrcSpace: ColorSpace>: ColorSpace {
    fn src_transform_raw(color: Vec3) -> Vec3;

    fn linear_part_raw(color: Vec3) -> Vec3;

    fn dst_transform_raw(color: Vec3) -> Vec3;
}

Performs the raw conversion from the ColorSpace represented by SrcSpc to the ColorSpace represented by Self in three concrete steps, each of which may do some work or be a no-op.

Required methods

fn src_transform_raw(color: Vec3) -> Vec3

fn linear_part_raw(color: Vec3) -> Vec3

fn dst_transform_raw(color: Vec3) -> Vec3