Trait colstodian::traits::ConvertToAlphaRaw

pub trait ConvertToAlphaRaw<DstAlphaState> {
    fn convert_raw(raw: Vec3, alpha: f32) -> Vec3;
}

The complement of ConvertFromAlphaRaw. Performs the conversion from alpha state Self into DstAlphaState on a raw color.

This is automatically implemented for all types that implement ConvertFromAlphaRaw, much like how the From and Into traits work, where From gets you Into for free.

Required methods

fn convert_raw(raw: Vec3, alpha: f32) -> Vec3