Trait colstodian::traits::ConvertToRaw

pub trait ConvertToRaw<DstSpace: ColorSpace>: ColorSpace {
    fn src_transform_raw(color: Vec3) -> Vec3;

    fn linear_part_raw(color: Vec3) -> Vec3;

    fn dst_transform_raw(color: Vec3) -> Vec3;
}

The complement of ConvertFromRaw.

Performs the raw conversion from the ColorSpace represented by Self to the ColorSpace represented by DstSpace in three concrete steps, each of which may do some work or be a no-op.

This is automatically implemented for all types that implement ConvertFromRaw, much like how the From and Into traits work, where From gets you Into for free.

Required methods

fn src_transform_raw(color: Vec3) -> Vec3

fn linear_part_raw(color: Vec3) -> Vec3

fn dst_transform_raw(color: Vec3) -> Vec3