Trait colstodian::traits::EncodedColorSpace

pub trait EncodedColorSpace: ColorSpace {
    type DecodedSpace: WorkingColorSpace + ConvertFromRaw<Self>;
}

Marks a type as representing an encoded color space.

A color in an EncodedColorSpace has few operations that can be performed on it other than just converting it to a WorkingColorSpace

Associated Types

type DecodedSpace: WorkingColorSpace + ConvertFromRaw<Self>