Trait colstodian::traits::WorkingColorSpace

pub trait WorkingColorSpace: ColorSpace { }

Marks a type as representing a color space that is not encoded and is therefore able to have many more operations performed on it.