Trait colstodian::ColorInto

pub trait ColorInto<DstCol> {
    // Required method
    fn color_into(self) -> DstCol;
}

Like Into but specialized for use with colstodian Color types.

This trait exists so that functions can accept colors in a variety of encodings generically in an ergonomic fashion. ColorInto is blanket implemented generically so that, if you have a function parameter impl ColorInto<Color<SomeEncoding>>, a Color in any other encoding that is able to .convert::<SomeEncoding>() can be passed into that function as argument directly.

See the docs of the convert method on Color for more.

Example

type MyColor = Color<LinearSrgb>;

fn test_fn(input: impl ColorInto<MyColor>) {
    let input: MyColor = input.color_into();
    let correct = Color::linear_srgb(0.14703, 0.42327, 0.22323);
    assert_eq_eps!(input, correct, 0.00001);
}

test_fn(Color::srgb_u8(107, 174, 130));
test_fn(Color::srgb_f32(0.41961, 0.68235, 0.5098));

Required Methods

fn color_into(self) -> DstCol